More fixes and validation

+++ crates/untanglers-core/src/optimizer_ops.rs
@@ -5,7 +5,7 @@ pub trait OptimizerOps<T>
 where
   T: Eq + Hash + Clone + Display + Debug,
 {
-  fn count_layer_crossings(&self, layer_index: usize) -> Result<i64, OptimizerError>;
+  fn count_layer_crossings(&self, layer_index: usize) -> Result<usize, OptimizerError>;
   fn count_crossings(&self) -> usize;
   fn get_nodes(&self) -> Vec<Vec<T>>;
 }
@@ -27,7 +27,7 @@ macro_rules! impl_optimizer_ops {
     where
       T: Eq + Hash + Clone + Display + Debug,
     {
-      fn count_layer_crossings(&self, layer_index: usize) -> Result<i64, OptimizerError> {
+      fn count_layer_crossings(&self, layer_index: usize) -> Result<usize, OptimizerError> {
         self.optimizer.count_layer_crossings(layer_index)
       }
       fn count_crossings(&self) -> usize {
